AgentsWithPhidata/requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
groq
openai
python-multipart
fastapi
uvicorn